all sampling and learning tests generalized

+++ src/spflow/torch/sampling/general/nodes/leaves/parametric/cond_geometric.py
@@ -65,6 +65,6 @@ def sample(
     sampling_ids = marg_ids & instance_ids_mask.bool().to(p.device)
 
     # data needs to be offset by +1 due to the different definitions between SciPy and PyTorch
-    data[sampling_ids, leaf.scope.query] = leaf.dist(p=p).sample((sampling_ids.sum(),)).to(p.device) + 1
+    data[sampling_ids, leaf.scope.query] = leaf.dist(p=p).sample((sampling_ids.sum(),)).double().to(p.device) + 1
 
     return data
